[1.0.x-rc] moving code of service layer from angular component .
