+++ main.js
@@ -16,11 +16,11 @@ function createWindow() {
   });
   //console.log(__dirname + '/src/favicon.ico');
   // and load the index.html of the app.
-  browserWindow.loadURL(`file://${__dirname}/dist/index.html`);
+  //browserWindow.loadURL(`file://${__dirname}/dist/index.html`);
 
   //browserWindow.setMenu(null);
 
-  //browserWindow.loadURL('http://localhost:4200');
+  browserWindow.loadURL('http://localhost:4200');
 
   // Open the DevTools.
   browserWindow.webContents.openDevTools();
