+++ src/app/module/vdemy.module.ts
@@ -23,8 +23,8 @@ import { AuthGuard } from '../auth/service/auth.guard';
 import { AuthService } from '../auth/service/auth.service';
 import { TimerTimeOutBroker } from '../auth/timeout/timeout-broker.service';
 import { PlayerDataService } from '../player/service/player-data.service';
-import { UnLoadGuard } from '../util/unload/service/unload.guard';
 import { SafeHtml } from '../util/pipe/safe-html.pipe';
+import { OnRouterNavigateGuard } from '../util/router/service/router-navigate.guard';
 
 const DECLRATATIONS_EXPORT_ARRAY = [
   AuthComponent,
@@ -56,7 +56,7 @@ const ENTRY_COMPONENTS_ARRAY = [EditCourseDialog, TimeoutDialog];
 const PROVIDERS_ARRAY = [
   AuthService,
   AuthGuard,
-  UnLoadGuard,
+  OnRouterNavigateGuard,
   DashboardService,
   TimeoutDialogService,
   TimerTimeOutBroker,
