+++ src/app/util/router/service/router-navigate.service.ts
@@ -1,0 +1,4 @@
+import { Observable } from 'rxjs';
+export interface OnRouterNavigate {
+  beforeRouterNavigate(): Observable<boolean> | Promise<boolean> | boolean;
+}
