--- src/app/util/unload/service/unload.service.ts
@@ -1,4 +1,0 @@
-import { Observable } from 'rxjs';
-export interface BeforeUnload {
-  beforeunload(): Observable<boolean> | Promise<boolean> | boolean;
-}
